Fix profiler command's "started" state desyncing
Fixed a rare case in the profiler commands making it unusable, if you entered a wrong argument, it would not start the profiler and tell you to enter it correctly, but next time you do so it will say it's already started, and if you try to stop it it would say it wasn't started to begin with. This behaviour is now corrected so that start command will not stay it's already started if the starting of the profiler got cancelled because of an argument error in the command.

The fix is to late init the SubCommand.profiler variable, since if an argument is wrong after initializing the profiler variable, it throws StopCodeflowException, which makes it never reach the .startProfiling() call, but the variable is still set, making the stop command think it is started. This creates both a leak of the Profiler instance and makes it unusable since both commands wouldn't work.