fix always showing spawning in 0s
if dragon hud is enabled but dragon spawning notification is disabled, it would always show spawning in 0s in the hud due to it not counting the spawn time if spawning notification is not enabled.

the fix counts the time if dragon spawning notification OR the dragon hud is enabled, however still only displays the spawning notification if only the spawning notification is enabled, so it will not display the spawning notification if just the hud enabled, but the spawn time will show correctly in the hud.