add sun gecko essence shop perk
and mention the (internal) perk name in the missing message, might be confusing if you are missing like +4 Strength from blessing of time perk and wither essence forbidden strength perk at the same time, you wouldn't know which is which.

didn't bother with a pretty name/display name instead of internal name for now as I'm planning to write a new front end display GUI for it in the future, as you can see in the current version stuff like UpgradeType and UpgradeSource are pretty much unused and not shown to the end user. in the future, we would use those to categorize upgrades and show them on different tabs or something.