update: changed drop chances and required meter XP of blaze slayer items
from latest skyblock update

Duplex 1 Book | 28,214 --> 23,220 | 1.7722 --> 2.1533
High Class Archfiend Dice | 230,240 --> 194,939 | 0.2172 --> 0.2565
Gabagool Distillate | 6,357 --> 10,649 | 7.8652 --> 4.6952

overall a buff, gabagool looks like a nerf but likely due to weight of items ingame it feels like we make more money from gabagool as well. im not sure if the old values of gabagool in the code are correct though as a drop rate of 7% would mean that before Magic Find wouldn't affect it, so im p sure it was lower than 5% before as well but i put the chance here wrong? perhaps it was because i took the values from wiki at the time but now i use the in-game rng meter