ensure all shutdown tasks run
if an exception occurs in one of the shutdown tasks, next shutdown tasks in the queue wouldn't run

also logs the exception properly indicating its an error in one of the shutdown tasks. sadly we can't have what task caused the error as a Runnable instance does not provide much information.