updating the display is a cheap operation, force update in any price change
formatNumber() would for example return 34M for both flat 34,000,000 coins, and 34,500,000 coins. The latter is enough price difference to change the Money/Hour value, but this if check checks if formatNumber for old price and new price is not same (which are same 34M for both 34,000,000 and 34,500,000) and prevents display from updating till you kill a boss, so just make the check to update in any price change.