add missing config checks and fix packet sent thread
adds config checks to not count frames and ticks if the respective features are not enabled.

fixes packet sent thread - previously, we were injecting to HEAD of addToSendQueue, which meant that our code would run on the thread queeing the packet and not in the actual netty thread when the packet is actually going to get dispatched. to fix it inject to calls of writeAndFlush instead. since one of the calls to that inside dispatchPacket method is called inside a Runnable (anonymous class implementation), we need to redirect the execute method and wrap the runnable.