remove armor stand optimizer bypass list
Entity#getCustomNameTag calls DataWatcher code that uses ReetrantReadWriteLock in each call, calling this for all Armor Stands is kind of expensive operation. After the recode, The Armor Stand Optimizer renders the closest 50 armor stands properly, so the bypass list is kind of useless now. Just remove it and we will think about it later if any need arises in the future.

NameTagCache was a class used to only call getCustomNameTag 1 time per the same entity (we need the name in multiple places, in Remove Blank Armor Stands mainly) - not removing that for now as it might still be useful in the future. We still call clearLastNameTag on it since removeIfBlankArmorStand calls the getCustomNameTag if motionX, motionY, motionZ are zero and the Armor Stand's inventory is empty. Since it does not call it unconditionally like us for all armor stands that does not hurt the perf as much.

The Sadan bypass list checking function is kept because that is still used inside the "Remove Armor Stands in Sadan & Wither King Fight" feature.