remove: main functionality of rbas
alternative features in the main feature are kept so no need to remove the feature from config entirely, but the main functionality is a maintainance burden for a little fps gain and was not worth keeping anymore. list of issues it caused so far ever since its addition:
 - removed relics in m7 causing you to unable to pickup them (fixed with a if condition to disable removing it before all relics are placed)
 - removed interaction armor stand in p3 to open terms causing you to unable to open any terminals (fixed with a if condition to disable removing when p3 starts)
 - removed the invisible armor stand you ride when hypixel is sending you between islands and when using a silkwire in the rift wyld woods to hang in the air, causing you to fall down instantly (fixed with a check that the ridingEntity of EntityPlayerSP is not the entity it's trying to remove)
 - removed the exit mineshaft text in mineshafts at the end portal at entrance to the shaft, causing you to unable to exit the mineshaft without using Ascension Rope (fixed with a ticksExisted check and isInMineshaft check and disabling removing if in mineshaft)

all of those were fixed with workaround checks in the past and served as a version update of darkaddons to address those issues, but there's now 2 new bugs discovered that we can't fix easily:
 - removes parkour time and social point leaderboard in private islands when you go far and then close again.
 - removes item names in the second bottom room of the auction house in the glass previews if you go far and then close again.
 - removes "Warp to Shen's Special Room" text when going far and close to the end portal entrance.

those issues used to not exist, and even with Armor Stand Optimizer and Remove Blank Armor Stands features both disabled, the nametags of those Armor Stands seem to get hidden when you move far and re-visible when you go close, however this a behaviour of Hypixel and not vanilla - getCustomNameTag.isEmpty() condition passes if you are not close, and does not if you are close since it has a name. ticksExisted does not reset to 0 when you go far and come close again since it's the same stand, just it's name being removed or re-added depending on your distance.

this behaviour wasn't like this when the feature was originally made, so I assume those bugs are new, but it would be hard to fix those bugs and the feature is just not worth maintaining anymore even if we could fix the 3 new bugs as the bugs will just keep happening and we would have too many hacky workaround if conditions.

the code is in the commit history in case it's ever needed again anyways, so it's direct remove and not commenting out.

The changes are justified for the reasons said earlier.