fix block pos optimizer crashing game on launch
unless -Xverify:none was specified, the transformer was calling the MethodMappings class to determine names of methods to transform which was calling PublicUtils.read to load the methods.v2.csv file, which was calling the Utils.read method - which loads the Utils class, which had references to MC classes so it was loading those as well, and this was causing the transformer to go boom crashing the game (affected versions: v0.3.0 - does not affect v0.2.0) - due to use of NBTTagCompound class in the newly added in v0.3.0 getItemLore method.

to strengthen against it, moved all MC related code to other classes: ItemUtils, ChunkUtils

to further make sure the issue does not happen again in the future, we have also moved the .read and .write methods to own class: FileUtils.