add: laggy server detector and solo crush helper
TBD: Thresholds for TPS for comment and coloring might be enhanced after further testing.
TBD: You need to pad when the counter shows 1 (since it's not instantenous due to nature of hypixel servers never being 20.0 tps) so that the crusher moves by the time Crush! is displayed at the moment, might need tweaking.